src/requirements.txt
boto3==1.9.151
Flask==1.0.3
python-dotenv==0.10.2
